roundabout/requirements.txt
# Python dependencies for Roundabout Simulation (Phase 1)
# Install with: pip install -r requirements.txt

# Core dependencies
pyyaml>=6.0          # Configuration file parsing
pandas>=1.5.0        # Data manipulation and analysis
numpy>=1.23.0        # Numerical operations

# Visualization
matplotlib>=3.6.0    # Static plotting
seaborn>=0.12.0      # Statistical visualizations
plotly>=5.11.0       # Interactive plots (optional but recommended)

# SUMO dependencies
# Note: SUMO itself must be installed separately
# See: https://sumo.dlr.de/docs/Installing/index.html
# The TraCI Python library is included with SUMO

# Optional: Enhanced functionality
scipy>=1.9.0         # Scientific computing (for advanced statistics)
